+++ src/main/java/io/nagurea/smsupsdk/accountmanaging/transfercredits/body/TransferBody.java
@@ -1,0 +1,16 @@
+package io.nagurea.smsupsdk.accountmanaging.transfercredits.body;
+
+import lombok.Builder;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.ToString;
+
+@Builder
+@ToString
+@Getter
+@EqualsAndHashCode
+public class TransferBody {
+
+    private final TransferInfo transfer;
+
+}